Extract shared HTML report shell out of render.ts
Moves the CSS, escapeHtml, formatBytes, and delta rendering into
report-shell.ts so the per-scenario report and the upcoming
cross-scenario summary cannot drift apart visually. Mechanical move,
no behavior change; render.vitest.ts's 150 tests still pass unchanged.
